Move IStreamletOperator and update comments
