+++ heron/api/src/java/org/apache/heron/streamlet/Streamlet.java
@@ -22,8 +22,6 @@ package org.apache.heron.streamlet;
 import java.util.List;
 
 import org.apache.heron.classification.InterfaceStability;
-import org.apache.heron.streamlet.impl.operators.IStreamletBasicOperator;
-import org.apache.heron.streamlet.impl.operators.IStreamletOperator;
 
 /**
  * A Streamlet is a (potentially unbounded) ordered collection of tuples.
