+++ heron/api/tests/java/org/apache/heron/streamlet/impl/StreamletImplTest.java
@@ -34,11 +34,11 @@ import org.apache.heron.resource.TestBolt;
 import org.apache.heron.resource.TestCustomOperator;
 import org.apache.heron.streamlet.Config;
 import org.apache.heron.streamlet.Context;
+import org.apache.heron.streamlet.IStreamletOperator;
 import org.apache.heron.streamlet.SerializableConsumer;
 import org.apache.heron.streamlet.SerializableTransformer;
 import org.apache.heron.streamlet.Streamlet;
 import org.apache.heron.streamlet.WindowConfig;
-import org.apache.heron.streamlet.impl.operators.IStreamletOperator;
 import org.apache.heron.streamlet.impl.streamlets.ConsumerStreamlet;
 import org.apache.heron.streamlet.impl.streamlets.CustomStreamlet;
 import org.apache.heron.streamlet.impl.streamlets.FilterStreamlet;
